+++ tfimm/architectures/__init__.py
@@ -1,6 +1,8 @@
 from .cait import *  # noqa: F401
 from .convmixer import *  # noqa: F401
 from .mlp_mixer import *  # noqa: F401
+from .pvt import *  # noqa: F401
+from .pvt_v2 import *  # noqa: F401
 from .resnet import *  # noqa: F401
 from .swin import *  # noqa: F401
 from .vit import *  # noqa: F401